Update JCheung-P4-README.md
Dear @peihongx and @jonathanvajda,

You are invited to review my Project 4. This is unfinished at the moment. I will let you know when it is good for further review. Thank you for your consideration, time, and for the review! Please let me know if you have any questions.

Best,
Jaron